perf: /dev/null useless properties
